Chore(Request Url) Read the query string of the url
What doesthis pr do?

The function passed into the http.createServer() has a req argument that represents the request from the client, as an object (http.IncomingMessage object).This object has a property called "url" which holds the part of the url that comes after the domain name:

Any related pivota stories?

[Finishes #164394579]